feat(dto): 更新ClusterRackDTO和AlertItem中的ID字段类型为Long，提升数据一致性
- 将ClusterRackDTO和AlertItem中的id和noticeGroupId字段类型从Integer更改为Long，以确保与数据模型的一致性。
- 更新ClusterRackVO中的id字段类型，确保系统在处理相关数据时的稳定性与兼容性。

这些改动增强了数据模型的健壮性与一致性，确保了整体设计的一致性。

+++ datasophon-common/src/main/java/com/datasophon/common/model/AlertItem.java
@@ -38,5 +38,5 @@ public class AlertItem implements Serializable {
 
     private Integer triggerDuration;
 
-    private Integer noticeGroupId;
+    private Long noticeGroupId;
 }
